fix(wayland): f64 wl_fixed ABI bug in pointer listener (wayland fix part 1)
wl_pointer_listener enter/motion surface_x/y + axis value were f64; the wire
type is wl_fixed_t (i32, 24.8). Int-vs-float register mismatch = garbage coords.
-> i32 in defines.rs + convert /256.0 in the events.rs handlers. linux-gnu clean.
First of the single-pass wayland marshalling repair (audit + spec in
WACOM_TOUCH_API_RESEARCH.md); marshalling wrappers next.

+++ dll/src/desktop/shell2/linux/wayland/defines.rs
@@ -220,8 +220,8 @@ pub struct wl_pointer_listener {
         pointer: *mut wl_pointer,
         serial: u32,
         surface: *mut wl_surface,
-        surface_x: f64,
-        surface_y: f64,
+        surface_x: i32, // wl_fixed_t (24.8); convert /256.0 in the handler
+        surface_y: i32,
     ),
     pub leave: extern "C" fn(
         data: *mut c_void,
@@ -233,8 +233,8 @@ pub struct wl_pointer_listener {
         data: *mut c_void,
         pointer: *mut wl_pointer,
         time: u32,
-        surface_x: f64,
-        surface_y: f64,
+        surface_x: i32, // wl_fixed_t (24.8); convert /256.0 in the handler
+        surface_y: i32,
     ),
     pub button: extern "C" fn(
         data: *mut c_void,
@@ -249,7 +249,7 @@ pub struct wl_pointer_listener {
         pointer: *mut wl_pointer,
         time: u32,
         axis: u32,
-        value: f64,
+        value: i32, // wl_fixed_t
     ),
     pub frame: extern "C" fn(data: *mut c_void, pointer: *mut wl_pointer),
     pub axis_source: extern "C" fn(data: *mut c_void, pointer: *mut wl_pointer, axis_source: u32),
